Fix lint and formatting issues
Starting on Flutter 2.5, Google established new standards lint rules
for Dart and Flutter (https://medium.com/dartlang/announcing-dart-2-14-b48b9bb2fb67).

This commit update the analysis_options.yaml and fix all code to meet
these guidelines.
