--- example/BoardItemObject.dart
@@ -1,11 +1,0 @@
-class BoardItemObject{
-
-  String? title;
-
-  BoardItemObject({this.title}){
-    if(this.title == null){
-      this.title = "";
-    }
-  }
-
-}
