--- example/BoardListObject.dart
@@ -1,16 +1,0 @@
-import 'BoardItemObject.dart';
-
-class BoardListObject{
-
-  String? title;
-  List<BoardItemObject>? items;
-
-  BoardListObject({this.title,this.items}){
-    if(this.title == null){
-      this.title = "";
-    }
-    if(this.items == null){
-      this.items = [];
-    }
-  }
-}
