+++ example/board_item_object.dart
@@ -1,0 +1,5 @@
+class BoardItemObject {
+  String? title;
+
+  BoardItemObject({this.title = ''});
+}
