+++ example/board_list_object.dart
@@ -1,0 +1,8 @@
+import 'board_item_object.dart';
+
+class BoardListObject {
+  String? title;
+  List<BoardItemObject>? items;
+
+  BoardListObject({this.title = '', this.items = const []});
+}
