+++ example/example.dart
@@ -1,11 +1,11 @@
 import 'package:boardview/board_item.dart';
 import 'package:boardview/board_list.dart';
-import 'package:boardview/boardview_controller.dart';
+import 'package:boardview/board_view_controller.dart';
 import 'package:flutter/material.dart';
-import 'package:boardview/boardview.dart';
+import 'package:boardview/board_view.dart';
 
-import 'BoardItemObject.dart';
-import 'BoardListObject.dart';
+import 'board_item_object.dart';
+import 'board_list_object.dart';
 
 class BoardViewExample extends StatelessWidget {
   final List<BoardListObject> _listData = [
@@ -14,9 +14,10 @@ class BoardViewExample extends StatelessWidget {
     BoardListObject(title: "List title 3")
   ];
 
-
   //Can be used to animate to different sections of the BoardView
-  final BoardViewController boardViewController = new BoardViewController();
+  final BoardViewController boardViewController = BoardViewController();
+
+  BoardViewExample({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -32,19 +33,17 @@ class BoardViewExample extends StatelessWidget {
 
   Widget buildBoardItem(BoardItemObject itemObject) {
     return BoardItem(
-        onStartDragItem: (int? listIndex, int? itemIndex, BoardItemState? state) {
-
-        },
+        onStartDragItem:
+            (int? listIndex, int? itemIndex, BoardItemState? state) {},
         onDropItem: (int? listIndex, int? itemIndex, int? oldListIndex,
             int? oldItemIndex, BoardItemState? state) {
           //Used to update our local item data
           var item = _listData[oldListIndex!].items![oldItemIndex!];
-          _listData[oldListIndex].items!.removeAt(oldItemIndex!);
+          _listData[oldListIndex].items!.removeAt(oldItemIndex);
           _listData[listIndex!].items!.insert(itemIndex!, item);
         },
-        onTapItem: (int? listIndex, int? itemIndex, BoardItemState? state) async {
-
-        },
+        onTapItem:
+            (int? listIndex, int? itemIndex, BoardItemState? state) async {},
         item: Card(
           child: Padding(
             padding: const EdgeInsets.all(8.0),
@@ -60,27 +59,23 @@ class BoardViewExample extends StatelessWidget {
     }
 
     return BoardList(
-      onStartDragList: (int? listIndex) {
-
-      },
-      onTapList: (int? listIndex) async {
-
-      },
+      onStartDragList: (int? listIndex) {},
+      onTapList: (int? listIndex) async {},
       onDropList: (int? listIndex, int? oldListIndex) {
         //Update our local list data
         var list = _listData[oldListIndex!];
-        _listData.removeAt(oldListIndex!);
+        _listData.removeAt(oldListIndex);
         _listData.insert(listIndex!, list);
       },
-      headerBackgroundColor: Color.fromARGB(255, 235, 236, 240),
-      backgroundColor: Color.fromARGB(255, 235, 236, 240),
+      headerBackgroundColor: const Color.fromARGB(255, 235, 236, 240),
+      backgroundColor: const Color.fromARGB(255, 235, 236, 240),
       header: [
         Expanded(
             child: Padding(
-                padding: EdgeInsets.all(5),
+                padding: const EdgeInsets.all(5),
                 child: Text(
                   list.title!,
-                  style: TextStyle(fontSize: 20),
+                  style: const TextStyle(fontSize: 20),
                 ))),
       ],
       items: items,
