+++ lib/board_item.dart
@@ -2,12 +2,14 @@ import 'package:boardview/board_list.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-typedef void OnDropItem(int? listIndex, int? itemIndex,int? oldListIndex,int? oldItemIndex, BoardItemState state);
-typedef void OnTapItem(int? listIndex, int? itemIndex, BoardItemState state);
-typedef void OnStartDragItem(
+typedef OnDropItem = void Function(int? listIndex, int? itemIndex,
+    int? oldListIndex, int? oldItemIndex, BoardItemState state);
+typedef OnTapItem = void Function(
     int? listIndex, int? itemIndex, BoardItemState state);
-typedef void OnDragItem(int oldListIndex, int oldItemIndex, int newListIndex,
-    int newItemIndex, BoardItemState state);
+typedef OnStartDragItem = void Function(
+    int? listIndex, int? itemIndex, BoardItemState state);
+typedef OnDragItem = void Function(int oldListIndex, int oldItemIndex,
+    int newListIndex, int newItemIndex, BoardItemState state);
 
 class BoardItem extends StatefulWidget {
   final BoardListState? boardList;
@@ -21,14 +23,14 @@ class BoardItem extends StatefulWidget {
 
   const BoardItem(
       {Key? key,
-        this.boardList,
-        this.item,
-        this.index,
-        this.onDropItem,
-        this.onTapItem,
-        this.onStartDragItem,
-        this.draggable = true,
-        this.onDragItem})
+      this.boardList,
+      this.item,
+      this.index,
+      this.onDropItem,
+      this.onTapItem,
+      this.onStartDragItem,
+      this.draggable = true,
+      this.onDragItem})
       : super(key: key);
 
   @override
@@ -37,7 +39,8 @@ class BoardItem extends StatefulWidget {
   }
 }
 
-class BoardItemState extends State<BoardItem> with AutomaticKeepAliveClientMixin{
+class BoardItemState extends State<BoardItem>
+    with AutomaticKeepAliveClientMixin {
   late double height;
   double? width;
 
@@ -46,20 +49,25 @@ class BoardItemState extends State<BoardItem> with AutomaticKeepAliveClientMixin
 
   void onDropItem(int? listIndex, int? itemIndex) {
     if (widget.onDropItem != null) {
-      widget.onDropItem!(listIndex, itemIndex,widget.boardList!.widget.boardView!.startListIndex,widget.boardList!.widget.boardView!.startItemIndex, this);
+      widget.onDropItem!(
+          listIndex,
+          itemIndex,
+          widget.boardList!.widget.boardView!.startListIndex,
+          widget.boardList!.widget.boardView!.startItemIndex,
+          this);
     }
     widget.boardList!.widget.boardView!.draggedItemIndex = null;
     widget.boardList!.widget.boardView!.draggedListIndex = null;
-    if(widget.boardList!.widget.boardView!.listStates[listIndex!].mounted) {
-      widget.boardList!.widget.boardView!.listStates[listIndex].setState(() { });
+    if (widget.boardList!.widget.boardView!.listStates[listIndex!].mounted) {
+      widget.boardList!.widget.boardView!.listStates[listIndex].setState(() {});
     }
   }
 
   void _startDrag(Widget item, BuildContext context) {
     if (widget.boardList!.widget.boardView != null) {
       widget.boardList!.widget.boardView!.onDropItem = onDropItem;
-      if(widget.boardList!.mounted) {
-        widget.boardList!.setState(() { });
+      if (widget.boardList!.mounted) {
+        widget.boardList!.setState(() {});
       }
       widget.boardList!.widget.boardView!.draggedItemIndex = widget.index;
       widget.boardList!.widget.boardView!.height = context.size!.height;
@@ -74,17 +82,17 @@ class BoardItemState extends State<BoardItem> with AutomaticKeepAliveClientMixin
             widget.boardList!.widget.index, widget.index, this);
       }
       widget.boardList!.widget.boardView!.run();
-      if(widget.boardList!.widget.boardView!.mounted) {
-        widget.boardList!.widget.boardView!.setState(() { });
+      if (widget.boardList!.widget.boardView!.mounted) {
+        widget.boardList!.widget.boardView!.setState(() {});
       }
     }
   }
 
   void afterFirstLayout(BuildContext context) {
-    try {
+    if (context.size != null) {
       height = context.size!.height;
       width = context.size!.width;
-    }catch(e){}
+    }
   }
 
   @override
@@ -98,10 +106,11 @@ class BoardItemState extends State<BoardItem> with AutomaticKeepAliveClientMixin
     widget.boardList!.itemStates.insert(widget.index!, this);
     return GestureDetector(
       onTapDown: (otd) {
-        if(widget.draggable) {
+        if (widget.draggable) {
           RenderBox object = context.findRenderObject() as RenderBox;
           Offset pos = object.localToGlobal(Offset.zero);
-          RenderBox box = widget.boardList!.context.findRenderObject() as RenderBox;
+          RenderBox box =
+              widget.boardList!.context.findRenderObject() as RenderBox;
           Offset listPos = box.localToGlobal(Offset.zero);
           widget.boardList!.widget.boardView!.leftListX = listPos.dx;
           widget.boardList!.widget.boardView!.topListY = listPos.dy;
@@ -124,7 +133,8 @@ class BoardItemState extends State<BoardItem> with AutomaticKeepAliveClientMixin
         }
       },
       onLongPress: () {
-        if(!widget.boardList!.widget.boardView!.widget.isSelecting && widget.draggable) {
+        if (!widget.boardList!.widget.boardView!.widget.isSelecting &&
+            widget.draggable) {
           _startDrag(widget, context);
         }
       },
