+++ lib/board_list.dart
@@ -1,11 +1,11 @@
 import 'package:boardview/board_item.dart';
-import 'package:boardview/boardview.dart';
+import 'package:boardview/board_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-typedef void OnDropList(int? listIndex,int? oldListIndex);
-typedef void OnTapList(int? listIndex);
-typedef void OnStartDragList(int? listIndex);
+typedef OnDropList = void Function(int? listIndex, int? oldListIndex);
+typedef OnTapList = void Function(int? listIndex);
+typedef OnStartDragList = void Function(int? listIndex);
 
 class BoardList extends StatefulWidget {
   final List<Widget>? header;
@@ -28,7 +28,10 @@ class BoardList extends StatefulWidget {
     this.headerBackgroundColor,
     this.boardView,
     this.draggable = true,
-    this.index, this.onDropList, this.onTapList, this.onStartDragList,
+    this.index,
+    this.onDropList,
+    this.onTapList,
+    this.onStartDragList,
   }) : super(key: key);
 
   final int? index;
@@ -39,25 +42,24 @@ class BoardList extends StatefulWidget {
   }
 }
 
-class BoardListState extends State<BoardList> with AutomaticKeepAliveClientMixin{
+class BoardListState extends State<BoardList>
+    with AutomaticKeepAliveClientMixin {
   List<BoardItemState> itemStates = [];
-  ScrollController boardListController = new ScrollController();
+  ScrollController boardListController = ScrollController();
 
   void onDropList(int? listIndex) {
-    if(widget.onDropList != null){
-      widget.onDropList!(listIndex,widget.boardView!.startListIndex);
+    if (widget.onDropList != null) {
+      widget.onDropList!(listIndex, widget.boardView!.startListIndex);
     }
     widget.boardView!.draggedListIndex = null;
-    if(widget.boardView!.mounted) {
-      widget.boardView!.setState(() {
-
-      });
+    if (widget.boardView!.mounted) {
+      widget.boardView!.setState(() {});
     }
   }
 
   void _startDrag(Widget item, BuildContext context) {
     if (widget.boardView != null && widget.draggable) {
-      if(widget.onStartDragList != null){
+      if (widget.onStartDragList != null) {
         widget.onStartDragList!(widget.index);
       }
       widget.boardView!.startListIndex = widget.index;
@@ -67,7 +69,7 @@ class BoardListState extends State<BoardList> with AutomaticKeepAliveClientMixin
       widget.boardView!.draggedItem = item;
       widget.boardView!.onDropList = onDropList;
       widget.boardView!.run();
-      if(widget.boardView!.mounted) {
+      if (widget.boardView!.mounted) {
         widget.boardView!.setState(() {});
       }
     }
@@ -82,13 +84,13 @@ class BoardListState extends State<BoardList> with AutomaticKeepAliveClientMixin
     List<Widget> listWidgets = [];
     if (widget.header != null) {
       listWidgets.add(GestureDetector(
-          onTap: (){
-            if(widget.onTapList != null){
+          onTap: () {
+            if (widget.onTapList != null) {
               widget.onTapList!(widget.index);
             }
           },
           onTapDown: (otd) {
-            if(widget.draggable) {
+            if (widget.draggable) {
               RenderBox object = context.findRenderObject() as RenderBox;
               Offset pos = object.localToGlobal(Offset.zero);
               widget.boardView!.initialX = pos.dx;
@@ -100,7 +102,7 @@ class BoardListState extends State<BoardList> with AutomaticKeepAliveClientMixin
           },
           onTapCancel: () {},
           onLongPress: () {
-            if(!widget.boardView!.widget.isSelecting && widget.draggable) {
+            if (!widget.boardView!.widget.isSelecting && widget.draggable) {
               _startDrag(widget, context);
             }
           },
@@ -112,51 +114,50 @@ class BoardListState extends State<BoardList> with AutomaticKeepAliveClientMixin
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: widget.header!),
           )));
-
     }
     if (widget.items != null) {
-      listWidgets.add(Container(
-          child: Flexible(
-              fit: FlexFit.loose,
-              child: new ListView.builder(
-                shrinkWrap: true,
-                physics: ClampingScrollPhysics(),
-                controller: boardListController,
-                itemCount: widget.items!.length,
-                itemBuilder: (ctx, index) {
-                  if (widget.items![index].boardList == null ||
-                      widget.items![index].index != index ||
-                      widget.items![index].boardList!.widget.index != widget.index ||
-                      widget.items![index].boardList != this) {
-                    widget.items![index] = new BoardItem(
-                      boardList: this,
-                      item: widget.items![index].item,
-                      draggable: widget.items![index].draggable,
-                      index: index,
-                      onDropItem: widget.items![index].onDropItem,
-                      onTapItem: widget.items![index].onTapItem,
-                      onDragItem: widget.items![index].onDragItem,
-                      onStartDragItem: widget.items![index].onStartDragItem,
-                    );
-                  }
-                  if (widget.boardView!.draggedItemIndex == index &&
-                      widget.boardView!.draggedListIndex == widget.index) {
-                    return Opacity(
-                      opacity: 0.0,
-                      child: widget.items![index],
-                    );
-                  } else {
-                    return widget.items![index];
-                  }
-                },
-              ))));
+      listWidgets.add(Flexible(
+          fit: FlexFit.loose,
+          child: ListView.builder(
+            shrinkWrap: true,
+            physics: const ClampingScrollPhysics(),
+            controller: boardListController,
+            itemCount: widget.items!.length,
+            itemBuilder: (ctx, index) {
+              if (widget.items![index].boardList == null ||
+                  widget.items![index].index != index ||
+                  widget.items![index].boardList!.widget.index !=
+                      widget.index ||
+                  widget.items![index].boardList != this) {
+                widget.items![index] = BoardItem(
+                  boardList: this,
+                  item: widget.items![index].item,
+                  draggable: widget.items![index].draggable,
+                  index: index,
+                  onDropItem: widget.items![index].onDropItem,
+                  onTapItem: widget.items![index].onTapItem,
+                  onDragItem: widget.items![index].onDragItem,
+                  onStartDragItem: widget.items![index].onStartDragItem,
+                );
+              }
+              if (widget.boardView!.draggedItemIndex == index &&
+                  widget.boardView!.draggedListIndex == widget.index) {
+                return Opacity(
+                  opacity: 0.0,
+                  child: widget.items![index],
+                );
+              } else {
+                return widget.items![index];
+              }
+            },
+          )));
     }
 
     if (widget.footer != null) {
       listWidgets.add(widget.footer!);
     }
 
-    Color? backgroundColor = Color.fromARGB(255, 255, 255, 255);
+    Color? backgroundColor = const Color.fromARGB(255, 255, 255, 255);
 
     if (widget.backgroundColor != null) {
       backgroundColor = widget.backgroundColor;
@@ -167,12 +168,12 @@ class BoardListState extends State<BoardList> with AutomaticKeepAliveClientMixin
     widget.boardView!.listStates.insert(widget.index!, this);
 
     return Container(
-        margin: EdgeInsets.all(8),
+        margin: const EdgeInsets.all(8),
         decoration: BoxDecoration(color: backgroundColor),
         child: Column(
           mainAxisSize: MainAxisSize.min,
           mainAxisAlignment: MainAxisAlignment.start,
-          children: listWidgets as List<Widget>,
+          children: listWidgets,
         ));
   }
 }
