+++ test/boardview_test.dart
@@ -1,9 +1,10 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:boardview/boardview.dart';
+// import 'package:boardview/boardview.dart';
 
 void main() {
   test('adds one to input values', () {
-    final boardview = BoardView();
+    // TODO: create tests
+    // final boardview = BoardView();
   });
 }
